+++ src/views/Drive/components/DriveExplorer/DriveExplorer.tsx
@@ -63,7 +63,6 @@ import { DriveTopBarItems } from './DriveTopBarItems';
 import { ShareDialogWrapper } from 'app/drive/components/ShareDialog/ShareDialogWrapper';
 import RealtimeService from 'services/sockets/socket.service';
 import { eventHandler } from 'services/sockets/event-handler.service';
-import VersionHistorySidebar from '../VersionHistorySidebar/VersionHistorySidebar';
 
 const MenuItemToGetSize = ({
   isTrash,
@@ -540,7 +539,6 @@ const DriveExplorer = (props: DriveExplorerProps): JSX.Element => {
       />
       <ClearTrashDialog onItemsDeleted={onTrashCleared} />
       <UploadItemsFailsDialog />
-      <VersionHistorySidebar />
       <MenuItemToGetSize isTrash={isTrash} translate={translate} menuItemsRef={menuItemsRef} />
       <ItemDetailsDialog onDetailsButtonClicked={onDetailsButtonClicked} />
       {editNameItem && (
