+++ src/views/Drive/components/VersionHistory/components/AutosaveSection.tsx
@@ -1,0 +1,36 @@
+import { Trash } from '@phosphor-icons/react';
+import { Checkbox } from '@internxt/ui';
+
+interface AutosaveSectionProps {
+  totalAutosaveCount: number;
+  selectAllAutosave: boolean;
+  onSelectAllChange: (checked: boolean) => void;
+  onDeleteAll: () => void;
+}
+
+export const AutosaveSection = ({
+  totalAutosaveCount,
+  selectAllAutosave,
+  onSelectAllChange,
+  onDeleteAll,
+}: AutosaveSectionProps) => {
+  if (totalAutosaveCount === 0) return null;
+
+  return (
+    <div className="group flex items-center justify-between border-b-[2.5px] border-gray-5 px-6 py-5 hover:bg-gray-1 dark:hover:bg-gray-5">
+      <div className="flex items-center space-x-3">
+        <Checkbox
+          checked={selectAllAutosave}
+          onClick={() => onSelectAllChange(!selectAllAutosave)}
+          className="h-4 w-4"
+        />
+        <span className="text-base text-gray-80">
+          {totalAutosaveCount}/{totalAutosaveCount} autosave versions
+        </span>
+      </div>
+      <button onClick={onDeleteAll} className="flex items-center justify-center">
+        <Trash size={24} className="text-[#FF0D0080]" />
+      </button>
+    </div>
+  );
+};
