+++ src/views/Drive/components/VersionHistory/components/index.ts
@@ -1,0 +1,4 @@
+export { Header } from './Header';
+export { CurrentVersionItem } from './CurrentVersionItem';
+export { VersionItem } from './VersionItem';
+export { AutosaveSection } from './AutosaveSection';
