+++ src/views/Drive/components/VersionHistory/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useDropdownPositioning } from './useDropdownPositioning';
+export { useVersionItemActions } from './useVersionItemActions';
