+++ src/views/Drive/components/VersionHistory/types.ts
@@ -1,0 +1,8 @@
+export interface FileVersion {
+  id: string;
+  date: Date;
+  userName: string;
+  expiresInDays?: number;
+  isAutosave?: boolean;
+  isCurrent?: boolean;
+}
